requirements.txt
PySide6
wiringpi
python-vlc
pyyaml
